feat: Write, test, and debug functionality to buy tokens

+++ programs/swings-dot-fun/src/utils/utils.rs
@@ -7,10 +7,11 @@ pub fn convert_str_to_pubkey(str: &str) -> Pubkey {
     Pubkey::from_str(str).unwrap()
 }
 
-pub fn calculate_amount_out(amount_in: &u64, reserve_in: &u64, reserve_out: &u64) -> u64 {
-    ((amount_in * reserve_out) as u128 / (amount_in + reserve_in) as u128) as u64
+pub fn calculate_amount_out(amount_in: &u128, reserve_in: &u128, reserve_out: &u128) -> u64 {
+    ((amount_in * reserve_out) / (amount_in + reserve_in)) as u64
 }
+// 50e9 * 100e9 / 50e9 + 100e9
 
-pub fn calculate_fee_amount(amount: &u64, trading_fee_in_bps: &u64) -> u64 {
-    ((amount * trading_fee_in_bps) as u128 / BPS as u128) as u64
+pub fn calculate_fee_amount(amount: &u128, trading_fee_in_bps: &u128) -> u64 {
+    ((amount * trading_fee_in_bps) / BPS as u128) as u64
 }
